requirements.txt
asgiref==3.11.1
Babel==2.17.0
Django==6.0.2
django-phonenumber-field==8.4.0
phonenumbers==9.0.24
pillow==12.1.1
shell==1.0.1
sqlparse==0.5.5
tzdata==2025.3
